+++ src/vue/pages/WillRequests/WillRequestsList.vue
@@ -14,7 +14,7 @@
         <template v-if="willRequests.length">
           <will-requests-table
             :will-requests="willRequests"
-            @submit="loadWillRequestsList"
+            @delete-will-request="$emit('delete-will-request', $event)"
           />
         </template>
         <template v-else>
@@ -48,22 +48,27 @@ export default {
 
   components: { Loader, LoadingErrorMessage, NoDataMessage, WillRequestsTable },
 
+  emits: ['delete-will-request'],
+
   setup () {
-    const store = useStore()
-    const accountId = computed(() => store.getters[vuexTypes.accountId])
-    const isAccountNotary =
-      computed(() => store.getters[vuexTypes.isAccountNotary])
-    const isAccountRegistry =
-      computed(() => store.getters[vuexTypes.isAccountRegistry])
-
-    const isNotaryOrRegistry =
-      computed(() => isAccountNotary.value || isAccountRegistry.value)
-
     const isLoading = ref(false)
     const isLoadFailed = ref(false)
     const willRequests = ref([])
 
-    Bus.on(Bus.eventList.createWillRequest, () => loadWillRequestsList())
+    const store = useStore()
+
+    const accountId = computed(() => store.getters[vuexTypes.accountId])
+    const isAccountNotary = computed(
+      () => store.getters[vuexTypes.isAccountNotary],
+    )
+    const isAccountRegistry = computed(
+      () => store.getters[vuexTypes.isAccountRegistry],
+    )
+    const isNotaryOrRegistry = computed(
+      () => isAccountNotary.value || isAccountRegistry.value,
+    )
+
+    Bus.on(Bus.eventList.willRequestManaged, () => loadWillRequestsList())
 
     const loadWillRequestsList = async () => {
       isLoading.value = true