+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.2] - 2021-06-15
+#### Added
+- Ability to general role account to delete will request on Will Requests and
+  Will Request Details pages
 #### Fixed
 - Filtering parameters for Notary and Registry roles in the `WillRequestList`
 - Restoring from `localStorage` after reloading page issue